refactor!: rename @sensejs/http to @sensejs/http-koa-platform
