+++ examples/demo-app/package.json
@@ -17,7 +17,8 @@
   "dependencies": {
     "@sensejs/config": "workspace:*",
     "@sensejs/core": "workspace:*",
-    "@sensejs/http": "workspace:*",
+    "@sensejs/http-common": "workspace:*",
+    "@sensejs/http-koa-platform": "workspace:*",
     "@sensejs/logger": "workspace:*",
     "@sensejs/typeorm": "workspace:*",
     "@sensejs/utility": "workspace:*",
