+++ examples/demo-app/src/http/example.controller.ts
@@ -1,4 +1,4 @@
-import {Body, Controller, GET, Path, POST, Query} from '@sensejs/http';
+import {Body, Controller, GET, Path, POST, Query} from '@sensejs/http-common';
 import {PublishingFacade} from '../example/publishing-facade.component';
 import {Inject, InjectLogger, Logger} from '@sensejs/core';
 import {validateCreateAuthorForm, validateCreateBookForm} from './http-validation';
